Support compile by XC32 ver 2.2, 1.42, and 1.40.

+++ z80.c
@@ -20,7 +20,7 @@
 	Local variables
 */
 
-volatile unsigned long g_timer1;
+volatile unsigned int g_timer1;
 
 static UINT16 z80PC;
 static UINT16 z80SP;
